version 1.0.6.0 PURE 0.1 by Analytica
When a ban is enforced on a new player, or a banned player trying to
join the server, it prints out a message in the server chat saying
"Enforcing ban on [player] for [reason]".

We'd like to disable this message output entirely. There aren't any
configuration options within the plugin on ProCon nor on the Metabans
website.